+++ src/auth/useAuthRole.js
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import AuthRoleContext from "./authRoleContext";
+
+export default function useAuthRole() {
+  return useContext(AuthRoleContext);
+}